Remove cast

+++ CPQKeys/external/nauty/dev_roanh_cpqkeys_algo_Nauty.c
@@ -1,5 +1,5 @@
 #include "dev_roanh_cpqkeys_algo_Nauty.h"
 
 JNIEXPORT jint JNICALL Java_dev_roanh_cpqkeys_algo_Nauty_test(JNIEnv* env, jclass obj, jint num){
-	return jint(2) * num;
+	return 2 * num;
 }
